+++ packages/@expo/metro-config/src/serializer/jsOutput.ts
@@ -5,6 +5,9 @@
  * LICENSE file in the root directory of this source tree.
  */
 import { FBSourceFunctionMap, MetroSourceMapSegmentTuple } from 'metro-source-map';
+import { JsTransformerConfig } from 'metro-transform-worker';
+
+import { Options as CollectDependenciesOptions } from '../transform-worker/collect-dependencies';
 
 export type JSFileType = 'js/script' | 'js/module' | 'js/module/asset';
 
@@ -21,6 +24,13 @@ export type JsOutput = {
       map: MetroSourceMapSegmentTuple[];
       functionMap: FBSourceFunctionMap | null;
     };
+
+    ast?: import('@babel/types').File;
+
+    hasCjsExports?: boolean;
+
+    readonly reconcile?: ReconcileTransformSettings;
+    readonly reactClientReference?: string;
   };
   type: JSFileType;
 };
@@ -40,6 +50,26 @@ export type ExpoJsOutput = Omit<JsOutput, 'data'> & {
       skipCache?: boolean;
     };
   };
+};
+
+export type ReconcileTransformSettings = {
+  inlineRequires: boolean;
+  importDefault: string;
+  importAll: string;
+  globalPrefix: string;
+  unstable_renameRequire?: boolean;
+  unstable_compactOutput?: boolean;
+  minify?: {
+    minifierPath: string;
+    minifierConfig: JsTransformerConfig['minifierConfig'];
+  };
+  collectDependenciesOptions?: CollectDependenciesOptions;
+
+  unstable_dependencyMapReservedName?: string;
+  optimizationSizeLimit?: number;
+  unstable_disableNormalizePseudoGlobals?: boolean;
+
+  normalizePseudoGlobals: boolean;
 };
 
 export function isExpoJsOutput(output: any): output is ExpoJsOutput {
